+++ BrainfuckInterceptor/Views/ConverterUI.xaml
@@ -1,0 +1,37 @@
+﻿<Page x:Class="BrainfuckInterceptor.Views.ConverterUI"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:BrainfuckInterceptor"
+      mc:Ignorable="d" 
+      d:DesignHeight="400" d:DesignWidth="430"
+      Title="ConverterUI" Background="{DynamicResource Background}" Foreground="{DynamicResource Foreground}">
+
+	<Grid>
+		<Grid.RowDefinitions>
+			<RowDefinition Height="30"/>
+			<RowDefinition Height="Auto"/>
+			<RowDefinition Height="Auto"/>
+			<RowDefinition Height="30"/>
+			<RowDefinition Height="Auto"/>
+			<RowDefinition Height="Auto"/>
+		</Grid.RowDefinitions>
+		<Grid.ColumnDefinitions>
+			<ColumnDefinition Width="10*"/>
+			<ColumnDefinition Width="100"/>
+			<ColumnDefinition Width="100"/>
+			<ColumnDefinition Width="10*"/>
+		</Grid.ColumnDefinitions>
+
+		<Label x:Name="InputLabel" Grid.Column="0" Grid.ColumnSpan="4" Grid.Row="0" Content="Input"/>
+		<TextBox x:Name="InputText" Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="4" Margin="0,0,0,10" Padding="5" Height="100" MinHeight="70"
+				AcceptsReturn="True" TextWrapping="Wrap" TextChanged="InputText_TextChanged"/>
+		<Button x:Name="EncodeButton" Grid.Column="1" Grid.Row="2" Content="Encode" HorizontalAlignment="Left" Width="90" MinHeight="40" Click="EncodeButton_Click" IsEnabled="False"/>
+		<Button x:Name="DecodeButton" Grid.Column="2" Grid.Row="2" Content="Decode" HorizontalAlignment="Right" Width="90" MinHeight="40" Click="DecodeButton_Click" IsEnabled="False"/>
+		<Label x:Name="OutputLabel" Grid.Column="0" Grid.ColumnSpan="4" Grid.Row="3" Content="Output"/>
+		<TextBox x:Name="OutputText" Grid.Column="0" Grid.Row="4" Grid.ColumnSpan="4" Margin="0,0,0,10" Padding="5" Height="100" MinHeight="70"
+				AcceptsReturn="True" TextWrapping="Wrap" IsReadOnly="True"/>
+		<Button x:Name="CopyButton" Grid.Column="1" Grid.Row="5" Grid.ColumnSpan="2" Content="Copy Output" HorizontalAlignment="Center" Width="120" MinHeight="40" Click="CopyButton_Click"/>
+	</Grid>
+</Page>
